+++ tests/casper/langs/test_binary_test_lang.py
@@ -1,7 +1,7 @@
 """The language testing module ... """
 import pytest
 
-from testing_languages.binary_test_lang import BinaryTestLang
+from state_languages.binary_test_lang import BinaryTestLang
 from casper.network import Network
 from casper.validator_set import ValidatorSet
