+++ tests/casper/langs/test_blockchain_test_lang.py
@@ -1,7 +1,7 @@
 """The language testing module ... """
 import pytest
 
-from testing_languages.blockchain_test_lang import BlockchainTestLang
+from state_languages.blockchain_test_lang import BlockchainTestLang
 from casper.network import Network
 from casper.validator_set import ValidatorSet
